Implemented flexible depth algorithm; minor tweaks
The new 'flexible depth' algorithm allows for much better control on the fragmentation. This means we can serve many more processes while they still finish collectively. The performance of this new algorithm (mainly in the master rank) is also much faster, and thus hardly a bottleneck on even short runs with many processes.